+++ src/clock.rs
@@ -1,14 +1,15 @@
 //! Abstraction for hardware timers/clocks
 
 use crate::{
-    duration::Duration, fixed_point::FixedPoint, fraction::Fraction, instant::Instant,
-    time_int::TimeInt, timer::param, timer::Timer,
+    duration::Duration, fraction::Fraction, instant::Instant, time_int::TimeInt, timer::param,
+    timer::Timer,
 };
 use core::hash::Hash;
 
 /// Potential `Clock` errors
 #[non_exhaustive]
 #[derive(Debug, Eq, PartialEq, Hash)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub enum Error {
     /// Exact cause of failure is unknown
     Unspecified,
@@ -34,7 +35,11 @@ impl Default for Error {
 /// software [`Timer`]s can be spawned from a `Clock` object.
 pub trait Clock: Sized {
     /// The type to hold the tick count
+    #[cfg(not(feature = "defmt"))]
     type T: TimeInt + Hash;
+    /// The type to hold the tick count
+    #[cfg(feature = "defmt")]
+    type T: TimeInt + Hash + defmt::Format;
 
     /// The duration of one clock tick in seconds, AKA the clock precision.
     const SCALING_FACTOR: Fraction;
@@ -51,10 +56,7 @@ pub trait Clock: Sized {
     fn new_timer<Dur: Duration>(
         &self,
         duration: Dur,
-    ) -> Timer<param::OneShot, param::Armed, Self, Dur>
-    where
-        Dur: FixedPoint,
-    {
+    ) -> Timer<param::OneShot, param::Armed, Self, Dur> {
         Timer::<param::None, param::None, Self, Dur>::new(&self, duration)
     }
 }