+++ src/duration.rs
@@ -11,6 +11,7 @@ use core::{
     convert::TryFrom,
     hash::{Hash, Hasher},
     mem::size_of,
+    ops,
     prelude::v1::*,
 };
 #[doc(hidden)]
@@ -290,7 +291,7 @@ pub use units::*;
 ///
 /// assert_eq!(Minutes(62_u32) % Hours(1_u32), Minutes(2_u32));
 /// ```
-pub trait Duration: Sized + Copy {
+pub trait Duration: FixedPoint + Sized + Copy {
     /// Construct a `Generic` `Duration` from a _named_ `Duration` (eg.
     /// [`Milliseconds`])
     ///
@@ -340,7 +341,6 @@ pub trait Duration: Sized + Copy {
         scaling_factor: Fraction,
     ) -> Result<Generic<DestInt>, ConversionError>
     where
-        Self: FixedPoint,
         DestInt: TryFrom<Self::T>,
     {
         Ok(Generic::<DestInt>::new(
@@ -395,8 +395,6 @@ pub trait Duration: Sized + Copy {
     /// ```
     fn to_rate<Rate: rate::Rate>(&self) -> Result<Rate, ConversionError>
     where
-        Rate: FixedPoint,
-        Self: FixedPoint,
         Rate::T: TryFrom<Self::T>,
     {
         let conversion_factor = Self::SCALING_FACTOR
@@ -406,11 +404,11 @@ pub trait Duration: Sized + Copy {
 
         if size_of::<Self::T>() >= size_of::<Rate::T>() {
             fixed_point::FixedPoint::from_ticks(
-                Self::T::from(*conversion_factor.numerator())
+                Self::T::from(conversion_factor.numerator())
                     .checked_div(
                         &self
                             .integer()
-                            .checked_mul(&Self::T::from(*conversion_factor.denominator()))
+                            .checked_mul(&Self::T::from(conversion_factor.denominator()))
                             .ok_or(ConversionError::Overflow)?,
                     )
                     .ok_or(ConversionError::DivByZero)?,
@@ -418,11 +416,11 @@ pub trait Duration: Sized + Copy {
             )
         } else {
             fixed_point::FixedPoint::from_ticks(
-                Rate::T::from(*conversion_factor.numerator())
+                Rate::T::from(conversion_factor.numerator())
                     .checked_div(
                         &Rate::T::try_from(self.integer())
                             .map_err(|_| ConversionError::Overflow)?
-                            .checked_mul(&Rate::T::from(*conversion_factor.denominator()))
+                            .checked_mul(&Rate::T::from(conversion_factor.denominator()))
                             .ok_or(ConversionError::Overflow)?,
                     )
                     .ok_or(ConversionError::DivByZero)?,
@@ -438,6 +436,7 @@ pub trait Duration: Sized + Copy {
 /// The purpose of this type is to allow a simple `Duration` object that can be defined at run-time.
 /// It does this by replacing the `const` _scaling factor_ with a struct field.
 #[derive(Copy, Clone, Debug, Default)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub struct Generic<T> {
     integer: T,
     scaling_factor: Fraction,
@@ -501,7 +500,108 @@ impl<T: TimeInt> Generic<T> {
     }
 }
 
-impl<T: TimeInt> Duration for Generic<T> {}
+impl<T: TimeInt> Generic<T> {
+    pub(crate) fn into_ticks<T2>(self, fraction: Fraction) -> Result<T2, ConversionError>
+    where
+        T2: TimeInt,
+        T2: TryFrom<T>,
+    {
+        if size_of::<T2>() > size_of::<T>() {
+            let ticks =
+                T2::try_from(self.integer()).map_err(|_| ConversionError::ConversionFailure)?;
+
+            if fraction > Fraction::new(1, 1) {
+                TimeInt::checked_div_fraction(
+                    &TimeInt::checked_mul_fraction(&ticks, &self.scaling_factor)
+                        .ok_or(ConversionError::Unspecified)?,
+                    &fraction,
+                )
+                .ok_or(ConversionError::Unspecified)
+            } else {
+                TimeInt::checked_mul_fraction(
+                    &ticks,
+                    &self
+                        .scaling_factor
+                        .checked_div(&fraction)
+                        .ok_or(ConversionError::Unspecified)?,
+                )
+                .ok_or(ConversionError::Unspecified)
+            }
+        } else {
+            let ticks = if self.scaling_factor > Fraction::new(1, 1) {
+                TimeInt::checked_div_fraction(
+                    &TimeInt::checked_mul_fraction(&self.integer(), &self.scaling_factor)
+                        .ok_or(ConversionError::Unspecified)?,
+                    &fraction,
+                )
+                .ok_or(ConversionError::Unspecified)?
+            } else {
+                TimeInt::checked_mul_fraction(
+                    &self.integer(),
+                    &self
+                        .scaling_factor
+                        .checked_div(&fraction)
+                        .ok_or(ConversionError::Unspecified)?,
+                )
+                .ok_or(ConversionError::Unspecified)?
+            };
+
+            T2::try_from(ticks).map_err(|_| ConversionError::ConversionFailure)
+        }
+    }
+
+    /// Checked addition of two `Generic` durations.
+    pub fn checked_add_generic<T2: TimeInt>(mut self, duration: Generic<T2>) -> Option<Self>
+    where
+        T: TryFrom<T2>,
+    {
+        let add_ticks: T = duration.into_ticks(*self.scaling_factor()).ok()?;
+        self.integer = self.integer.checked_add(&add_ticks)?;
+
+        Some(self)
+    }
+
+    /// Checked subtraction of two `Generic` durations.
+    pub fn checked_sub_generic<T2: TimeInt>(mut self, duration: Generic<T2>) -> Option<Self>
+    where
+        T: TryFrom<T2>,
+    {
+        let sub_ticks: T = duration.into_ticks(*self.scaling_factor()).ok()?;
+        self.integer = self.integer.checked_sub(&sub_ticks)?;
+
+        Some(self)
+    }
+}
+
+impl<T: TimeInt, T2: TimeInt> ops::Add<Generic<T2>> for Generic<T>
+where
+    T: TryFrom<T2>,
+{
+    type Output = Self;
+
+    fn add(self, rhs: Generic<T2>) -> Self::Output {
+        if let Some(v) = self.checked_add_generic(rhs) {
+            v
+        } else {
+            panic!("Add failed")
+        }
+    }
+}
+
+impl<T: TimeInt, T2: TimeInt> ops::Sub<Generic<T2>> for Generic<T>
+where
+    T: TryFrom<T2>,
+{
+    type Output = Self;
+
+    fn sub(self, rhs: Generic<T2>) -> Self::Output {
+        if let Some(v) = self.checked_sub_generic(rhs) {
+            v
+        } else {
+            panic!("Sub failed")
+        }
+    }
+}
 
 /// Duration units
 #[doc(hidden)]
@@ -560,9 +660,15 @@ pub mod units {
                 }
             }
 
+            #[cfg(feature = "defmt")]
+            impl<T: TimeInt + defmt::Format> defmt::Format for $name<T> {
+                fn format(&self, fmt: defmt::Formatter) {
+                    defmt::write!(fmt, "{}", self.0)
+                }
+            }
+
             impl<T: TimeInt, Rhs: Duration> ops::Add<Rhs> for $name<T>
             where
-                Rhs: FixedPoint,
                 Self: TryFrom<Rhs>,
             {
                 type Output = Self;
@@ -576,7 +682,6 @@ pub mod units {
             impl<T: TimeInt, Rhs: Duration> ops::Sub<Rhs> for $name<T>
             where
                 Self: TryFrom<Rhs>,
-                Rhs: FixedPoint,
             {
                 type Output = Self;
 
@@ -623,7 +728,6 @@ pub mod units {
             impl<T: TimeInt, Rhs: Duration> ops::Rem<Rhs> for $name<T>
             where
                 Self: TryFrom<Rhs>,
-                Rhs: FixedPoint,
             {
                 type Output = Self;