+++ src/lib.rs
@@ -272,7 +272,7 @@
 //! - [`RTIC`](https://github.com/rtic-rs/cortex-m-rtic)
 //! - [`time`](https://docs.rs/time/latest/time) (Specifically the [`time::NumbericalDuration`](https://docs.rs/time/latest/time/trait.NumericalDuration.html)
 //!   implementations for primitive integers)
-#![doc(html_root_url = "https://docs.rs/embedded-time/0.12.0")]
+#![doc(html_root_url = "https://docs.rs/embedded-time/0.13.0")]
 #![deny(unsafe_code)]
 #![cfg_attr(not(test), no_std)]
 #![warn(missing_docs)]
@@ -295,6 +295,7 @@ pub use timer::Timer;
 /// Crate errors
 #[non_exhaustive]
 #[derive(Debug, Eq, PartialEq, Hash)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub enum TimeError {
     /// Exact cause of failure is unknown
     Unspecified,
@@ -325,6 +326,7 @@ impl Default for TimeError {
 /// Conversion errors
 #[non_exhaustive]
 #[derive(Debug, Eq, PartialEq, Hash)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub enum ConversionError {
     /// Exact cause of failure is unknown
     Unspecified,